+++ fixtures/direct/src/client.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import {
   getSyncMessage,
@@ -74,4 +74,4 @@ const node = document.createElement('div');
 
 document.body.appendChild(node);
 
-render(<App />, node);
+createRoot(node).render(<App />);