+++ fixtures/server/webpack.config.js
@@ -24,7 +24,7 @@ module.exports = () => [
           exclude: [/node_modules/],
           use: [
             {
-              loader: 'babel-loader',
+              loader: require.resolve('babel-loader'),
               options: {
                 babelrc: false,
                 plugins: ['@loadable/babel-plugin'],
@@ -66,7 +66,7 @@ module.exports = () => [
           exclude: [/node_modules/],
           use: [
             {
-              loader: 'babel-loader',
+              loader: require.resolve('babel-loader'),
               options: {
                 babelrc: false,
                 plugins: ['@loadable/babel-plugin'],