+++ fixtures/simple/src/client.tsx
@@ -1,6 +1,6 @@
 import { VocabProvider, useTranslations } from '@vocab/react';
 import React, { type ReactNode, useState } from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import commonTranslations from './.vocab';
 import clientTranslations from './client.vocab';
@@ -67,9 +67,8 @@ const node = document.createElement('div');
 
 document.body.appendChild(node);
 
-render(
+createRoot(node).render(
   <App>
     <Content />
   </App>,
-  node,
 );